+++ src/8thWonderland/Library/Application.php
@@ -4,6 +4,8 @@ namespace Wonderland\Library;
 
 use Pimple\Container;
 
+use Wonderland\Library\Http\Request\Request;
+
 class Application {
     /** @var string **/
     protected $rootPath;
@@ -157,6 +159,11 @@ class Application {
         }
     }
     
+    /**
+     * @param string $controllerName
+     * @param string $actionName
+     * @throws \Exception
+     */
     public function startControllerAction($controllerName, $actionName) {
         $controller = "Wonderland\\Application\\Controller\\{$controllerName}Controller";
         $action = "{$actionName}Action";
@@ -166,6 +173,8 @@ class Application {
         if (!method_exists($controller, $action)) {
             throw new \Exception("The Action '$action' does not exist !");
         }
-        (new $controller($this))->$action();
+        $response = (new $controller($this, new Request()))->$action();
+        $response->makeHeaders();
+        $response->respond();
     }
 }